+++ app.js
@@ -336,7 +336,7 @@ global.string = function (str) {
 
 global.LoginServer = require('./loginserver.js');
 
-global.Wulu = require('./wulu/index');
+global.Wulu = require('./wulu');
 
 // Connect to MongoDB.
 Wulu.Mongo.connect_database();